#163 - tweaks - Defined Types better, Make Chronology email notification take place at save only, Make things consistent added interface for legacy notification service, Added NotificationTimeFrame base class to use between the two different notification services class to share code and avoid duplication, Activated isOnHours functionaly for new Notification service class
