+++ src/main/java/us/com/plattrk/api/model/EntityType.java
@@ -1,6 +1,6 @@
 package us.com.plattrk.api.model;
 
-public enum Type {
+public enum EntityType {
 
     INCIDENT,
     RESOLUTION,
